[Chore] #39 - 타입 변경에 따른 코드 수정

+++ src/main/java/org/moonshot/server/global/common/model/validator/TargetNumberValidator.java
@@ -4,10 +4,10 @@ import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 import lombok.extern.slf4j.Slf4j;
 
-public class TargetNumberValidator implements ConstraintValidator<ValidTargetNumber, Integer> {
+public class TargetNumberValidator implements ConstraintValidator<ValidTargetNumber, Long> {
 
     @Override
-    public boolean isValid(Integer value, ConstraintValidatorContext context) {
+    public boolean isValid(Long value, ConstraintValidatorContext context) {
         if (value == null) {
             return true;
         } else {
